+++ src/main/java/ru/yandex/practicum/filmorate/service/event/impl/EventServiceImpl.java
@@ -1,13 +1,13 @@
-package ru.yandex.practicum.filmorate.service.film.impl;
+package ru.yandex.practicum.filmorate.service.event.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import ru.yandex.practicum.filmorate.dao.EventDao;
+import ru.yandex.practicum.filmorate.dao.event.EventDao;
 import ru.yandex.practicum.filmorate.exeption.UserServiceException;
 import ru.yandex.practicum.filmorate.model.Event;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.service.film.EventService;
-import ru.yandex.practicum.filmorate.storage.user.UserStorage;
+import ru.yandex.practicum.filmorate.service.event.EventService;
+import ru.yandex.practicum.filmorate.dao.user.UserDao;
 
 import java.util.List;
 
@@ -17,7 +17,7 @@ import static java.lang.String.format;
 @RequiredArgsConstructor
 public class EventServiceImpl implements EventService {
     private final EventDao eventDao;
-    private final UserStorage userStorage;
+    private final UserDao userDao;
 
     @Override
     public List<Event> getUserFeed(int userId) {
@@ -26,7 +26,7 @@ public class EventServiceImpl implements EventService {
     }
 
     private User checkUser(Integer userId) {
-        return userStorage.getUserById(userId)
+        return userDao.getUserById(userId)
                 .orElseThrow(() -> new UserServiceException(
                         format("Попытка запросить ленту пользователя с несуществующим id: %d", userId)));
     }
